Multithreaded Construction Heuristics
