+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/heuristic/thread/MoveEvaluationOperation.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.optaplanner.core.impl.heuristic.thread;
+
+import org.optaplanner.core.impl.heuristic.move.Move;
+
+public class MoveEvaluationOperation<Solution_> extends MoveThreadOperation<Solution_> {
+
+    private final int stepIndex;
+    private final int moveIndex;
+    private final Move<Solution_> move;
+
+    public MoveEvaluationOperation(int stepIndex, int moveIndex, Move<Solution_> move) {
+        this.stepIndex = stepIndex;
+        this.moveIndex = moveIndex;
+        this.move = move;
+    }
+
+    public int getStepIndex() {
+        return stepIndex;
+    }
+
+    public int getMoveIndex() {
+        return moveIndex;
+    }
+
+    public Move<Solution_> getMove() {
+        return move;
+    }
+
+}
